split up test cases and distributed them amongst multiple files
